Split big file into small ones
